requirements.txt
kafka-python
spacy==3.0.9
setuptools>=67.8.0
wheel>=0.40.0
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_ner_bc5cdr_md-0.4.0.tar.gz
pyyaml
confluent-kafka
confluent-kafka[avro]
fastavro